+++ tools/run-stat/num-gen.c
@@ -55,9 +55,9 @@ static p_ng_node ng_make_node( const uint64_t start, const uint64_t count )
 }
 
 
-static int CC ng_insert_helper( const void* item1, const void* item2 )
+static int64_t CC ng_insert_helper( const void* item1, const void* item2 )
 {
-    int res = 0;
+    int64_t res = 0;
     p_ng_node node1 = (p_ng_node)item1;
     p_ng_node node2 = (p_ng_node)item2;
     if ( node1->start < node2->start )
